feat(upgrade): change rewards multiplier for mainnet (#472)
Add to change reward multiplier of period delegation for Mainnet. 

For Aeneid, the Virgil upgrade has already executed, so this change of
upgrade handler do not change the multipliers for Aeneid. That means,
anyone can sync blocks from the genesis with the new binary.

This will only change the multipliers for Mainnet (`chainID` =
`story-1`). This assumes that no period delegations will happen until
the upgrade height since it will be in the singularity.

Tested with localnet for the 2 cases
- mocking Aeneid (already upgrade executed, and sync block from the
genesis with this new binary)
- mocking Mainnet (change staking params including rewards multiplier)

issue: none