feat(app): add with comet flag (#482)
Added with-comet flag, which is a flag for running an app with CometBFT.
The default value is `true`.
This is to support KYVE's state sync, which uses external CometBFT
process that feeds state-sync snapshots.

Also some changes in CLI flags for running node are made per KYVE team's
request. With this changes, customized pruning is supported with
`--pruning custom` flag.

## Added flags

- `with-comet`: Run abci app embedded in-process with CometBFT
- `address`: Address of proxy app
- `transport`: Specify abci transport (socket | grpc)
- `pruning-keep-recent`: Number of recent heights to keep on disk
(ignored if pruning is not 'custom')
- `pruning-interval`: Height interval at which pruned heights are
removed from disk (ignored if pruning is not 'custom'), this is not used
by this command but kept for compatibility with the complete pruning
options

## Updated

Some flags are renamed. These would lead to breaking changes in CLI
usage for operation.

- rename flags of `--log-color`, `--log-format` and `--log-level` to
`--log_color`, `--log_format` and `--log_level`
- rename flags of `--snapshot-interval` and `--snapshot-keep-recent` to
`--state-sync.snapshot-interval` and `--state-sync.snapshot-keep-recent`

issue: #487