fix(cli): fix validator not found err in validator creation (#515)
This PR is based on #513.

When creating validator using CLI, if the validator doesn't exist,
story-api returned error with http code of 500. It leads to that
`isValidatorFound` return error, so couldn't create a validator even if
the validator under the pubkey does not exist. Handle the case of
validator not found appropriately.

issue: none