fix(evmstaking): handle residual reward by version (#569)
In Story v1.2.1, the handling of residual rewards was upgraded. As a
result, when syncing from the genesis block using a client running
v1.2.1 or later, a state difference may occur at a certain block,
leading to an app hash mismatch error. To address this issue, the
residual reward handling logic has been corrected to properly account
for the block height — applying the appropriate logic before and after
the v1.2.1 upgrade point.

issue: none