feat(evmengine): defer upgrade scheduling to upgrade keeper (#566)
This PR adds support for rolling upgrades via the `UpgradeEntrypoint`
contract. The core implementation is mostly adapted from the `signal`
module in Celestia. Instead of introducing a new module, the logic is
integrated into the `evmengine` module, which is responsible for
handling EVM-triggered upgrades. Since our upgrades are initiated from
within the EVM and do not require validator quorum, we were able to
adopt only the necessary parts of the `signal` module and implement them
within `evmengine` for simplicity.

**Functionally, this is very similar to how we previously supported
rolling upgrades using `Fork`.**

As you may know, in Cosmos SDK, a node must not upgrade its binary
before the scheduled upgrade height, if the upgrade has already been
registered in the `upgrade` keeper. This constraint prevents validators
from upgrading early when the upgrade is pre-scheduled.

To address this, we can defer scheduling upgrade. When the
`UpgradeEntrypoint` contract emits an upgrade event, the upgrade
information is temporarily stored in the `evmengine`'s state. The actual
scheduling of the upgrade in the `upgrade` keeper is deferred until the
block reaches the specified upgrade height. At that point, during the
`PreBlock` phase, the upgrade is scheduled in `upgrade` keeper so that
it can be applied within that same block.

issue: #149